Adjust project template to current feature set (#40)
## 📥 Pull Request Description

The configurations (`yaml` files), part of the project template, are
incorrect or incomplete. In addition, the TensorFlow version stored in
the project template is still wrong (`2.8`).

With this pull request the configurations of the template were adjusted
and the TensorFlow version was changed to `2.9`.

## 👀 Affected Areas

- Project template (`./template`) 

## 📝 Checklist

Please make sure you've completed the following tasks before submitting
this pull request:

- [X] Pre-commit hooks were executed
- [ ] Changes have been reviewed by at least one other developer
- [ ] ~~Tests have been added or updated to cover the changes (only
necessary if the changes affect the executable code)~~(**Not
necessary**)
- [ ] ~~Documentation has been updated to reflect the changes~~ (**Not
necessary**)
- [X] Any necessary migrations have been run

## 📌 Related Issues

Closes #35 , closes #36 

## 🔗 Links

Please provide any relevant links (e.g. documentation, external
resources) that support your changes.

## 📷 Screenshots

If applicable, please include screenshots of the before and after
effects of your changes.