feat: Add abstract `FileChecksumProcessor` and concrete class `ZippedCsvToParquetProcessor` (#44)
## 📥 Pull Request Description

With a FileChecksumProcessors, files from an input directory can be
processed and the results can then be saved in an output directory. The
checksums of input and output files are created and saved so that only
files that have changed are processed when executed again.

## 👀 Affected Areas

Core implementation. Mostly part of a sort of data preprocessing
pipeline.

## 📝 Checklist

Please make sure you've completed the following tasks before submitting
this pull request:

- [X] Pre-commit hooks were executed
- [ ] Changes have been reviewed by at least one other developer
- [X] Tests have been added or updated to cover the changes (only
necessary if the changes affect the executable code)
- [ ] Documentation has been updated to reflect the changes
- [X] Any necessary migrations have been run

## 📌 Related Issues

*None*

## 🔗 Links

*None*

## 📷 Screenshots

*None*

---------

Co-authored-by: Denis Stalz-John <denis.stalz-john@codecentric.de>