fix: Remove temp directory from hydra search path. Add hydra config mapping factory (#47)
## 📥 Pull Request Description

Until now, a temporary directory was always added to the search path of
a `dagster` configuration. This caused `dagster` to give an error
message when the configurations had to be searched.
The temporary directory is necessary for the operation with `dagit`,
because no configuration file has to be specified there, but directly
YAML code. In this pull request the temporary directory was removed from
the searchpath configuration.
Also, there is now a `hydra_conf_mapping_factory` that is used as a
decorator of a Dagster job, as this allows parameters to be passed and
overridden.

## 👀 Affected Areas

- Pipelines 

## 📝 Checklist

Please make sure you've completed the following tasks before submitting
this pull request:

- [X] Pre-commit hooks were executed
- [ ] Changes have been reviewed by at least one other developer
- [X] Tests have been added or updated to cover the changes (only
necessary if the changes affect the executable code)
- [ ] Documentation has been updated to reflect the changes - **Not
necessary**
- [X] Any necessary migrations have been run 
- Pipelines run via a `dagster` script, but can also be started via
`dagit`.

## 📌 Related Issues

- closes #42 

## 🔗 Links

*None*

## 📷 Screenshots

`dagster` pipeline run

<img width="1586" alt="Bildschirmfoto 2023-07-05 um 08 08 35"
src="https://github.com/codecentric-oss/niceml/assets/48205130/661a1cf7-7c01-45ed-bbdf-5f87abd9df26">


`dagit` pipeline run 


![Capture-2023-07-05-080913](https://github.com/codecentric-oss/niceml/assets/48205130/785da3f9-eaf5-45eb-9bfb-8666317675dd)