fix: Optimize file locks and remove outdated template config (#72)
## 📥 Pull Request Description

The 'is_acquired' parameter of FileLocks can now be set when the Lock is
initialized. This can be necessary, if a FileLock file is already
present, but the according FileLock object is not. This may happen, if a
FileLock could not have been released e.g. due to an error.

Additionally, an outdated config was removed from a template
configuration.

## 👀 Affected Areas

Everywhere a FileLock is used.

## 📝 Checklist

Please make sure you've completed the following tasks before submitting
this pull request:

- [x] Pre-commit hooks were executed
- [x] Changes have been reviewed by at least one other developer
- [ ] Tests have been added or updated to cover the changes (only
necessary if the changes affect the executable code)
- [x] All tests ran successfully
- [x] All merge conflicts are resolved
- [ ] Documentation has been updated to reflect the changes
- [ ] Any necessary migrations have been run